Removed router, fixes style, architecture.
