+++ frontend/src/routes.js
@@ -3,14 +3,13 @@ import main from './components/main';
 import mainLayout from './components/layout/mainLayout';
 import eoiHeader from './components/eois/eoiHeader';
 import overview from './components/eois/overview';
-import overviewProfile from './components/organizationProfile/overview';
 import pinned from './components/eois/pinned';
 import calls from './components/eois/calls';
 import direct from './components/eois/direct';
 import partner from './components/partners/partnersHeader';
 import dashboard from './components/dashboard/dashboard';
 import applications from './components/applications/applications';
-import profileHeader from './components/organizationProfile/organizationProfileHeader';
+import organizationProfile from './components/organizationProfile/organizationProfile';
 import settings from './components/agencySettings/agencySettings';
 import registration from './components/registration/registration';
 import mainContent from './components/common/mainContentWrapper';
@@ -41,17 +40,7 @@ export default [
         },
         { path: 'partner', component: partner },
         { path: 'applications', component: applications },
-        {
-          path: 'profile',
-          component: profileHeader,
-          childRoutes: [
-            { component: mainContent,
-              childRoutes: [
-                { path: 'overview', component: overviewProfile },
-              ],
-            },
-          ],
-        },
+        { path: 'profile', component: organizationProfile },
         { path: 'settings', component: settings },
       ],
     }],
